abstraction refactor
