+++ app/repository/characterRepository.go
@@ -1,45 +1,193 @@
 package repository
 
 import (
+	"encoding/csv"
+	"encoding/json"
 	"errors"
+	"fmt"
+	"io"
+	"io/ioutil"
+	"math"
+	"net/http"
+	"os"
+	"strconv"
+	"sync"
 
 	"go-project/app/domain"
 )
 
-type CharactersRepo struct{}
+const CSVFile string = "/Users/alejandrosanchez/Documents/go_bootcamp/app/resources/characters.csv"
+const URL string = "https://swapi.dev/api/people"
+
+var counter = 0
+
+type csvI interface {
+	ReadCsvFiletoString(filePath string) ([][]string, error)
+	Addrows(characters domain.Characters, filePath string) error
+}
 
 type Response struct {
 	Results domain.Characters `json:"results"`
 }
 
+type CharactersRepo struct {
+	csvC csvI
+	file string
+	api  string
+}
+
 //NewCharacterRepo factory Character repo
-func NewCharacterRepo() CharactersRepo {
-	return CharactersRepo{}
+func NewCharacterRepo(csvC csvI, file string, api string) CharactersRepo {
+	return CharactersRepo{csvC, file, api}
 }
 
 //FindAll gets characters from csv
-func (CharactersRepo) FindAll(data [][]string) (domain.Characters, error) {
+func (cr CharactersRepo) FindAll() (domain.Characters, error) {
+	data, err := cr.csvC.ReadCsvFiletoString(CSVFile)
+
+	if err != nil {
+		return nil, errors.New(err.Error())
+	}
+
 	characterList := domain.CreateCharacterList(data)
 
 	return characterList, nil
 }
 
-//FindByID get character in the csv by ID
-func (CharactersRepo) FindByID(data [][]string, characterID int) (domain.Character, error) {
-	characterList := domain.CreateCharacterList(data)
+//FetchCharacters gets characters from an API
+func (cr CharactersRepo) FetchCharacters() (domain.Characters, error) {
+	response, err := http.Get(URL)
+	if err != nil {
+		return nil, errors.New(err.Error())
+	}
 
-	character := domain.Character{}
+	responseData, err := ioutil.ReadAll(response.Body)
 
-	for _, v := range characterList {
-		if v.ID == characterID {
-			character = v
-			break
+	if err != nil {
+		return nil, errors.New(err.Error())
+	}
+
+	responseObject := Response{}
+	json.Unmarshal(responseData, &responseObject)
+
+	fmt.Println(responseObject.Results)
+
+	data, err := cr.csvC.ReadCsvFiletoString(CSVFile)
+
+	fmt.Println(data)
+
+	if err != nil {
+		return nil, errors.New(err.Error())
+	}
+
+	lastId := len(data)
+
+	fmt.Println(lastId)
+
+	for i := 0; i < len(responseObject.Results); i++ {
+		responseObject.Results[i].ID = lastId + 1
+		lastId++
+	}
+
+	cr.csvC.Addrows(responseObject.Results, CSVFile)
+
+	charactersCsv, err := cr.FindAll()
+
+	if err != nil {
+		return nil, errors.New(err.Error())
+	}
+
+	charactersCsv = append(charactersCsv, responseObject.Results...)
+
+	return charactersCsv, nil
+}
+
+// WorkerPoolCsv csv reader worker pool
+func (cr CharactersRepo) WorkerPoolCsv(t string, items int, itpw int) (domain.Characters, error) {
+	file, err := os.Open(CSVFile)
+
+	if err != nil {
+		return nil, errors.New(err.Error())
+	}
+
+	defer file.Close()
+
+	reader := csv.NewReader(file)
+
+	var characters domain.Characters
+
+	workers := int(math.Ceil(float64(items) / float64(itpw)))
+
+	jobs := make(chan []string)
+	res := make(chan domain.Character, items-1)
+
+	var wg sync.WaitGroup
+
+	for w := 0; w < workers; w++ {
+		wg.Add(1)
+
+		go func() {
+			defer wg.Done()
+			cr.worker(t, jobs, res, items)
+		}()
+	}
+
+	go func() {
+		for {
+			record, err := reader.Read()
+			if err == io.EOF {
+				break
+			} else if err != nil {
+				continue
+			}
+			jobs <- record
 		}
+		close(jobs)
+	}()
+
+	go func() {
+		wg.Wait()
+		close(res)
+	}()
+
+	for r := range res {
+		characters = append(characters, r)
 	}
 
-	if (domain.Character{}) == character {
-		return domain.Character{}, errors.New("error: character doesn't exist")
-	}
+	counter = 0
 
-	return character, nil
+	return characters, nil
+}
+
+// worker worker from WorkerPoolCsv
+func (CharactersRepo) worker(t string, jobs <-chan []string, results chan domain.Character, items int) {
+	for {
+		select {
+		case job, ok := <-jobs:
+			if !ok {
+				return
+			}
+
+			if counter == items {
+				return
+			}
+
+			id, _ := strconv.Atoi(job[0])
+
+			if t == "odd" && id%2 == 0 {
+				continue
+			} else if t == "even" && id%2 != 0 {
+				continue
+			}
+
+			character, err := domain.CreateCharacter(job)
+
+			if err != nil {
+				continue
+			}
+
+			results <- character
+		}
+		counter++
+	}
 }
